Uploading initial files
login.php - Page for student login. Has a redirect link for registration if the user doesn't have an account yet. Include a link to admin/tutor login page (and registration on that page too). Idea is then they'll be sent to the admin page

register.php - Page of student registration. Has a redirect link for login if the user already has an account.

loginRegBackend.php - Backend for login and register pages.

test.php - Placeholder for where the student gets sent after login. To include (add payment info, page to schedule appointments (or just do it there)).